+++ custom_components/podwaffle/const.py
@@ -6,7 +6,7 @@ from typing import Final, TypedDict
 from homeassistant.const import Platform
 
 DOMAIN: Final = "podwaffle"
-INTEGRATION_VERSION: Final = "0.3.0"
+INTEGRATION_VERSION: Final = "0.4.0"
 
 CONF_BASE_URL: Final = "base_url"
 CONF_JOIN_CODE: Final = "join_code"
@@ -14,7 +14,12 @@ CONF_PROFILE_IDS: Final = "profile_ids"
 CONF_PROFILES: Final = "profiles"
 CONF_VERIFY_SSL: Final = "verify_ssl"
 
-PLATFORMS: Final = [Platform.MEDIA_PLAYER, Platform.NOTIFY, Platform.SENSOR]
+PLATFORMS: Final = [
+    Platform.BUTTON,
+    Platform.MEDIA_PLAYER,
+    Platform.SENSOR,
+    Platform.TEXT,
+]
 UPDATE_INTERVAL: Final = timedelta(seconds=60)
 STATS_INTERVAL: Final = timedelta(minutes=5)